Minor advance on subject reduction
Problem:

It seams that it is necessary to merge the reduction, expansion and
subtype rule into one rule:

    Gamma |- t : T
    Gamma |- U : s
    T <= U
    --------------
    Gamma |- t : U

Reason: If the reduction, expansion and subtype rule remain separated,
we have to prove that all intermediate types are valid types. This is in
some cases not possible. Therefore the expansion rule has already been
strengthened to the reflexive transitive closure of beta reduction.

	modified:   src/ecc/ecc_typing.tex