Changed the definition of the typing relation
The structural rule for beta reduction is not sufficient to prove the
subject reduction lemma. It is necessary to state in the definition of
the typing relation that a beta equivalent type as long as it is a valid
type of some sort is a valid type for the same term.

	modified:   src/cc/cc_typing.tex