Tried a failing proof
I have tried to proof that all one step reductions of a minimal
diverging term remain diverging. This is not true. I have found a
counterexample which reduces a minimal diverging term to a strongly
normalizing term.

Problem: The one step reduction might throw away arguments which are one
of the causes of divergence e.g. a term of the form 'x x' which can be
replaced by 'omega omega'. If the one step reduction throws away 'x x',
then the term can become strongly normalizing.

	modified:   src/ecc/ecc_macros.tex
	modified:   src/ecc/ecc_thoughts.tex